INT-9554: Postgresql compatibility changes to queries
Issues addressed:

* Unsubscribe all action wouldn't work:
DELETE statements cannot use inner joins, the solution posted here:
https://github.com/moodlerooms/moodle-mod_hsuforum/pull/4/commits/8119124766e07a58c927c4690edccb3007e8f23a
was already in use in EM since august 2014.

* The "View posters" page would return a sql exception:
psql requires that a select statement include all the fields used in a
group by statement - grouping by the primary key of a table is enough to
then let you perform a select on any field in that table however (since
psql 9.1).

* Viewing a discussion would return an sql exception:
psql doesn't let you perform an ORDER BY operation on a field without it
being included in the SELECTed fields. Adding it to the selected fields
would result in multiple entries for the same user in the result set
(which would then violate the unique first column requirement).