INT-9966: mod_hsuforum fix timed discussion creation
The error was "java.util.HashMap cannot be cast to
java.lang.String (WebDriver\Exception\UnknownError)"

This was due to using the wrong method to select
the value in a select form element.

Also changed from ceil to floor to prevent it from
trying to select "60" in the menu which does not
exist.